Windows port: pool: fix MSVC narrowing conversions and always_inline macro
- device/kernels/fixed_point_arithmetic.hpp: ALWI macro used GNU
  __attribute__((always_inline)); this header is pulled into host TU's
  (rotate program factories), so give it an MSVC branch using __forceinline.
- generic/device/pool_multi_core_program_factory.cpp: static_cast loop index
  and DeviceAddr values down to the narrower field/arg types (uint16_t
  ScalarInfo fields, uint32_t Shape dims, uint32_t compile-time args).
- generic/generic_pools.cpp: static_cast int pad_top/bottom/left/right to
  uint32_t for the std::array<uint32_t, 4> aggregate return.
- upsample/device/upsample_common.cpp, upsample_program_factory_multicore_sharded.cpp,
  upsample/upsample.cpp: static_cast float/size_t/int expressions down to
  uint32_t at brace-init sites (shard shape, config Shape dims,
  SlidingWindowConfig::scale_h/scale_w).

+++ ttnn/cpp/ttnn/operations/pool/upsample/device/upsample_program_factory_multicore_sharded.cpp
@@ -205,7 +205,8 @@ Tensor create_config_tensor(
         config_vector.size(),
         elems_per_core);
 
-    ttnn::Shape config_shape({tt::div_up(config_vector.size(), elems_per_core), elems_per_core});
+    ttnn::Shape config_shape(
+        {static_cast<uint32_t>(tt::div_up(config_vector.size(), elems_per_core)), elems_per_core});
     auto config_buffer = HostBuffer(std::move(config_vector));
     return Tensor(std::move(config_buffer), config_shape, DataType::UINT16, Layout::ROW_MAJOR);
 }
